Create TileCostCalculator.java
Very basic interpretation of this project. It asks for a width and length of area to be tiled and a cost per unit of tile. It then calculates the area of the tiled area, assuming the area is rectangular, and gives the total cost of the tile. Because the concept of the project is simple enough, I used this to reintroduce some GUI concepts to my java programming.